requirement.txt
streamlit
langchain
langchain-community
langchain-google-community
langchain-openai
langgraph
openai
python-dotenv
pandas
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client


